api/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
granian = { version = "^1.4.1", extras = ["reload"] }
SQLAlchemy = "^2.0.30"
psycopg = { extras = ["binary"], version = "^3.1.19" }
notebook = "^7.2.0"
pgvector = "^0.2.5"
celery = "^5.4.0"
redis = { version = "^5.0.5", extras = ["hiredis"] }
gevent = "^23.9.1"
pandas = "^2.2.2"
fastapi-users = { extras = ["sqlalchemy"], version = "^13.0.0" }
pydantic-settings = "^2.3.3"
alembic = "^1.13.1"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
